% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_finalise.R
\name{clean_data_finalise}
\alias{clean_data_finalise}
\title{Finalise data cleaning}
\usage{
clean_data_finalise(data, test, prolific_filter, prolific_filter_label)
}
\value{

}
\description{
Filter out relevant prolific participants, remove duplicates, and add column
ID
}
\details{
Only remove duplicates now because sometimes the past study filtering is not
set right in Prolific and you still need to credit the duplicates

Assign IDs only after all filtering

Filter out testing data and data generated through Prolific preview (through
personal PID)
}
\author{
Shir Dekel
}
