% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation4.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{subject}{}
\item{condition}{}
\item{presentation}{}
\item{distribution}{}
\item{awareness}{}
\item{proportion}{}
\item{portfolio_binary}{}
\item{sex}{}
\item{age}{}
\item{language}{}
\item{language_other}{}
\item{business_edu}{}
\item{business_exp}{}
\item{current}{}
\item{sector}{}
\item{sector_other}{}
\item{employees}{}
\item{revenue}{}
\item{role_company}{}
\item{role_company_other}{}
\item{role_allocation}{}
\item{role_allocation_other}{}
\item{budget}{}
\item{project_expectation}{}
\item{project_number}{}
\item{portfolio_number}{}
\item{stage}{}
\item{experiment}{}
\item{sample}{}
\item{similarity}{}
\item{thesis_project}{}
\item{project_order}{}
\item{project}{}
\item{detail}{}
\item{outcome_positive}{}
\item{outcome_dif}{}
\item{probability_positive}{}
\item{choice}{}
\item{datetime}{}
\item{total_time}{}
\item{gamble}{}
\item{expected_value}{}
\item{gain_loss_ratio}{}
\item{date_rank}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{aggregation4}
}
