% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_materials_experiment4.R
\name{get_file_name_materials_experiment4}
\alias{get_file_name_materials_experiment4}
\title{Get E4 screenshot file names}
\usage{
get_file_name_materials_experiment4()
}
\value{

}
\description{
Get E4 screenshot file names
}
\author{
Shir Dekel
}
