% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions_experiment4.R
\name{get_instructions_experiment4}
\alias{get_instructions_experiment4}
\title{Get instructions for E4}
\usage{
get_instructions_experiment4()
}
\value{

}
\description{
Get instructions for E4
}
\author{
Shir Dekel
}
