% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_components_experiment3.R
\name{get_project_components_experiment3}
\alias{get_project_components_experiment3}
\title{Get details for long projects

Names generated from namelix.com}
\usage{
get_project_components_experiment3()
}
\value{

}
\description{
Get details for long projects

Names generated from namelix.com
}
\author{
Shir Dekel
}
