% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_description_experiment3.R
\name{get_project_description_experiment3}
\alias{get_project_description_experiment3}
\title{Get long project descriptions}
\usage{
get_project_description_experiment3(
  gambles,
  project_components_experiment3,
  project_detail_experiment3
)
}
\arguments{
\item{project_long_detail}{}
}
\value{

}
\description{
Get long project descriptions
}
\author{
Shir Dekel
}
