% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_detail_components_input_experiment3.R
\name{get_project_detail_components_input_experiment3}
\alias{get_project_detail_components_input_experiment3}
\title{Get input-friendly components}
\usage{
get_project_detail_components_input_experiment3(
  project_detail_components_experiment3
)
}
\arguments{
\item{project_detail_components_experiment3}{}
}
\value{

}
\description{
Get input-friendly components
}
\author{
Shir Dekel
}
