% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_detail_experiment3.R
\name{get_project_detail_experiment3}
\alias{get_project_detail_experiment3}
\title{Get detail line}
\usage{
get_project_detail_experiment3(project_detail_components_experiment3)
}
\arguments{
\item{project_detail_components_experiment3}{}
}
\value{

}
\description{
Get detail line
}
\author{
Shir Dekel
}
