% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_pair_sample.R
\name{get_project_pair_sample}
\alias{get_project_pair_sample}
\title{Get project pair sample}
\usage{
get_project_pair_sample(detail)
}
\value{

}
\description{
Get project pair sample
}
\author{
Shir Dekel
}
