% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_sections_experiment3.R
\name{get_project_sections_experiment3}
\alias{get_project_sections_experiment3}
\title{Get project sections E3}
\usage{
get_project_sections_experiment3(gambles, components, detail)
}
\value{

}
\description{
Get project sections E3
}
\author{
Shir Dekel
}
