% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_project_expectation.R
\name{get_trial_project_expectation}
\alias{get_trial_project_expectation}
\title{Project expectation follow up question}
\usage{
get_trial_project_expectation()
}
\value{

}
\description{
Project expectation follow up question
}
\author{
Shir Dekel
}
