% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_separate_experiment4.R
\name{get_trial_separate_experiment4}
\alias{get_trial_separate_experiment4}
\title{Get separate condition timeline for E4}
\usage{
get_trial_separate_experiment4(
  preamble,
  timeline_variable,
  form_options = c("Yes", "No"),
  awareness,
  project_variation_value,
  randomize_order = randomize_order
)
}
\arguments{
\item{preamble}{}

\item{project_input}{}
}
\value{

}
\description{
Get separate condition timeline for E4
}
\author{
Shir Dekel
}
