% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trials_experiment4.R
\name{get_trials_experiment4}
\alias{get_trials_experiment4}
\title{Get long projects distribution absent trial}
\usage{
get_trials_experiment4(projects_experiment4, randomize_order)
}
\arguments{
\item{projects_experiment4}{}

\item{randomize_order}{Each participant will see 20 project made up of the 10 domains, with two variations each. The two variations will be chosen at random, such that there will be 10 sets of pair variations.}
}
\value{

}
\description{
Get long projects distribution absent trial
}
\author{
Shir Dekel
}
