% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_project_sections_experiment3.R
\name{shuffle_project_sections_experiment3}
\alias{shuffle_project_sections_experiment3}
\title{Shuffle E3 project sections in JS}
\usage{
shuffle_project_sections_experiment3(project_sections_experiment3)
}
\arguments{
\item{project_sections_experiment3}{}
}
\value{

}
\description{
Shuffle E3 project sections in JS
}
\author{
Shir Dekel
}
