% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation4.R
\docType{package}
\name{aggregation4}
\alias{aggregation4}
\alias{aggregation4-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
aggregation4
A data package for aggregation4.

Use \code{data(package='aggregation4')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
